requirements.txt
pywebview>=5.0
pyautogui>=0.9
Pillow>=10.0
pywin32>=306
numpy>=1.24
psutil>=5.9
